pyproject.toml
[project]
name = "papersift"
version = "0.1.0"
description = "Entity-based paper clustering for Claude Code"
requires-python = ">=3.11"
dependencies = [
    "igraph>=0.10",
    "leidenalg>=0.10",
    "numpy>=1.20",
    "scikit-learn>=1.0",
]

[project.optional-dependencies]
enrich = ["pyalex>=0.14", "tqdm>=4.0"]
ui = [
    "dash>=2.17.0",
    "dash-cytoscape>=1.0.2",
    "dash-ag-grid>=33.3.0",
    "pandas>=2.0.0",
    "plotly>=5.0.0",
    "networkx>=3.0",
]
landscape = ["umap-learn>=0.5"]
all = ["papersift[enrich,ui,landscape]"]

[project.scripts]
papersift = "papersift.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/papersift"]
